+++ packages/liveblocks-core/src/index.ts
@@ -83,7 +83,12 @@ export {
 } from "./immutable";
 export { kInternal } from "./internal";
 export { assert, assertNever, nn } from "./lib/assert";
-export type { AsyncResult, AsyncResultWithDataField } from "./lib/AsyncResult";
+export type {
+  AsyncError,
+  AsyncLoading,
+  AsyncResult,
+  AsyncSuccess,
+} from "./lib/AsyncResult";
 export { autoRetry } from "./lib/autoRetry";
 export { chunk } from "./lib/chunk";
 export {
@@ -111,6 +116,8 @@ export { asPos, makePosition } from "./lib/position";
 export type { Resolve } from "./lib/Resolve";
 export { shallow } from "./lib/shallow";
 export { stringify } from "./lib/stringify";
+export type { QueryParams, URLSafeString } from "./lib/url";
+export { url, urljoin } from "./lib/url";
 export type { Brand, DistributiveOmit } from "./lib/utils";
 export {
   b64decode,